requirements.txt
faiss-cpu
futures
langchain
langchain-openai
matplotlib
networkx
nltk
numpy
python-dotenv
scikit-learn
spacy
tqdm
langchain-community
rank-bm25
PyMuPDF
deepeval
pypdf
Flask
llama-cloud-services
